- fixed the jxr large-image encode fix so it builds on modern GCC/Clang
76cb194 widened the 32-bit `cbStride * i` overflow in the JXR encoder by
casting through `unsigned long long`, but assigned the result to `void*
wmiBI.pv` without a cast. That is a C constraint violation: MSVC only
warns, while GCC >= 14 and Clang >= 15 make -Wint-conversion an error, so
JXRGlueJxr.c:833 and :957 were the only two lines in the whole tree that
failed to compile.

Keep pointer arithmetic and widen the offset with (size_t) instead. This
preserves the >4 GiB fix (a 40000x36000 24bpp frame wrapped at row 35792,
landing at offset 72704 instead of 4295040000) and is byte-identical to
the previous code below 4 GiB: A/B against the pre-fix object produced
identical output for all 28 encodes tested (24 synthetic across 8/24/32bpp
and pitch pads 0-3, plus the 4 HDR RGBAF samples), and all 24 synthetic
JXR_LOSSLESS round-trips are pixel-exact.

Also drop the now-unused `bpp` left behind in PluginJXR.cpp by the revert
in 8c723a6.

+++ Source/LibJXR/jxrgluelib/JXRGlueJxr.c
@@ -830,7 +830,7 @@ ERR PKImageEncode_EncodeContent_Encode(
 		Bool f420 = ( pIE->WMP.wmiI.cfColorFormat == YUV_420 || 
 					  (pIE->WMP.wmiSCP.bYUVData && pIE->WMP.wmiSCP.cfColorFormat==YUV_420) );
         CWMImageBufferInfo wmiBI = { 0 };
-        wmiBI.pv = (unsigned long long)pbPixels + (unsigned long long)cbStride * i / (f420 ? 2 : 1);
+        wmiBI.pv = pbPixels + (size_t)cbStride * i / (f420 ? 2 : 1);
         wmiBI.cLine = min(16, cLine - i);
         wmiBI.cbStride = cbStride;
         FailIf(ICERR_OK != ImageStrEncEncode(pIE->WMP.ctxSC, &wmiBI), WMP_errFail);
@@ -954,7 +954,7 @@ ERR PKImageEncode_EncodeAlpha_Encode(
     for (i = 0; i < cLine; i += 16)
     {
         CWMImageBufferInfo wmiBI = { 0 };
-        wmiBI.pv = (unsigned long long)pbPixels + (unsigned long long)cbStride * i;
+        wmiBI.pv = pbPixels + (size_t)cbStride * i;
         wmiBI.cLine = min(16, cLine - i);
         wmiBI.cbStride = cbStride;
         FailIf(ICERR_OK != ImageStrEncEncode(pIE->WMP.ctxSC_Alpha, &wmiBI), WMP_errFail);
